+++ WarringStates/Flow/SpanFlowTickOnArgs.cs
@@ -2,9 +2,9 @@
 
 namespace WarringStates.Flow;
 
-public class SpanFlowTickOnArgs(int currentSpan, Date currentDate) : ICallbackArgs
+public class SpanFlowTickOnArgs(long currentSpan, Date currentDate) : ICallbackArgs
 {
-    public int CurrentSpan { get; } = currentSpan;
+    public long CurrentSpan { get; } = currentSpan;
 
     public Date CurrentDate { get; } = currentDate;
 }
